+++ TombEngine/Objects/TR4/Trap/SpikyCeiling.h
