+++ TombEngine/Objects/TR4/tr4_objects.cpp
@@ -55,6 +55,7 @@
 // Switches
 
 // Traps
+#include "Objects/TR4/Trap/SpikyCeiling.h"
 #include "Objects/TR4/Trap/tr4_birdblade.h"
 #include "Objects/TR4/Trap/tr4_blade.h"
 #include "Objects/TR4/Trap/tr4_catwalkblade.h"
@@ -70,7 +71,6 @@
 #include "Objects/TR4/Trap/tr4_slicerdicer.h"
 #include "Objects/TR4/Trap/tr4_spikeball.h"
 #include "Objects/TR4/Trap/tr4_spikywall.h"
-#include "Objects/TR4/Trap/tr4_spikyceiling.h"
 #include "Objects/TR4/Trap/tr4_stargate.h"
 #include "Objects/TR4/Trap/tr4_cog.h"
 #include "Objects/TR4/Trap/tr4_teethspike.h"
