+++ TombEngine/TombEngine.vcxproj
@@ -569,6 +569,7 @@ xcopy /Y "$(SolutionDir)Libs\zlib\x64\*.dll" "$(TargetDir)"</Command>
     <ClInclude Include="Objects\TR4\Object\tr4_sarcophagus.h" />
     <ClInclude Include="Objects\TR4\Object\tr4_scales.h" />
     <ClInclude Include="Objects\TR4\Object\tr4_senet.h" />
+    <ClInclude Include="Objects\TR4\Trap\SpikyCeiling.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_birdblade.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_blade.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_catwalkblade.h" />
@@ -584,7 +585,6 @@ xcopy /Y "$(SolutionDir)Libs\zlib\x64\*.dll" "$(TargetDir)"</Command>
     <ClInclude Include="Objects\TR4\Trap\tr4_sethblade.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_slicerdicer.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_spikeball.h" />
-    <ClInclude Include="Objects\TR4\Trap\tr4_spikyceiling.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_spikywall.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_stargate.h" />
     <ClInclude Include="Objects\TR4\Trap\tr4_teethspike.h" />
@@ -1011,6 +1011,7 @@ xcopy /Y "$(SolutionDir)Libs\zlib\x64\*.dll" "$(TargetDir)"</Command>
     <ClCompile Include="Objects\TR4\Object\tr4_senet.cpp" />
     <ClCompile Include="Objects\TR4\Object\WraithTrap.cpp" />
     <ClCompile Include="Objects\TR4\tr4_objects.cpp" />
+    <ClCompile Include="Objects\TR4\Trap\SpikyCeiling.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_birdblade.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_blade.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_catwalkblade.cpp" />
@@ -1026,7 +1027,6 @@ xcopy /Y "$(SolutionDir)Libs\zlib\x64\*.dll" "$(TargetDir)"</Command>
     <ClCompile Include="Objects\TR4\Trap\tr4_sethblade.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_slicerdicer.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_spikeball.cpp" />
-    <ClCompile Include="Objects\TR4\Trap\tr4_spikyceiling.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_spikywall.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_stargate.cpp" />
     <ClCompile Include="Objects\TR4\Trap\tr4_teethspike.cpp" />
